Add a from_string constructor and a method to get bytes as a string
Being able to create a new BSSize instance from a string and to return the
number of bytes as a string is the crucial first step allowing the
implementation to support sizes bigger than what fits into an unsigned 64bit
integer.

Also fix some minor issues in the basic implementation.