Also try translated units when parsing a size string
People may enter sizes in their native format so we need to try such units
too. We should also check just the first n characters of units where n is length
of the unit string. This makes things like "10 G" work. Last but not least we
should ignore the case of the unit string because it doesn't really matter.