Use both GMP and MPFR libraries to correctly parse numeric parts
I know this looks stupid, but MPFR doesn't handle things like 1٫5 (valid number
in the ps_AF.UTF-8 locale using a unicode radix character) and on the other
hand, GMP thinks 0.1*1000 = 99. So until MPFR gets things right (because in case
of GMP it's not a bug but a feature) we need to use both of them to parse a
single numeric string or do our own complex magic.