Make sure our tests don't get broken by installed overrides
If overrides are installed for some GI-loaded module/library, GI loads them. So
if we run tests on our local sources/build and overrides from the installed
package are somewhere in the system they would get loaded and blast our
tests. We need to prevent this from happening and there seems to be no nice way
to do so.

So much for defending the hack I'm adding here.