Make sure we return Size if doing operations with big integers
If a given integer is bigger than MAXUINT64, we need to use a trick and convert
it to a Size instance first. But then the result of the operation is an int or
float because we are dividing/multiplying/... Size by Size. Thus we need to
convert the result to Size because that's what Size/int is supposed to return.