Ditch autopoint.
This removes the need for all of that junk in the po directory. In
particular, LINGUAS is no longer needed, and can no longer get out of
sync with the translations that are actually available.

libbytesize.pot is no longer kept in source control or distributed in
the tarball. It will be generated during a build and before making a
release.

Removed the po-empty target since it is no longer needed. Moved the
translation-canary tests from po-pull to dist-hook so that bad .po files
are excluded from the tarball but kept in the source directory. This
makes it easier to see what's actually going wrong if something is going
wrong.