Make sure Size can be interpreted as integer in Python 3.10
Having __int__ is no longer enough to have Size interpreted as
integer. Starting with 3.10 we need to implement __index__ for
that because __int__ is not considered lossless and is also
available for floats and similar "non-integer" numbers where it is
used for lossy conversions (rounding).
See https://bugs.python.org/issue37999 for more information.